Ingest capability demonstration.
Python pmetrics library created with demo of webserver, along with a
probably-useful node agent to serve data.

Conveyance script uses Named Pipes, curl, and a little bash to plug
the node agent to the piper.

(this should be substantially faster than having the conveyer download
to an actual file and *then* to write to an HTTP api - it also
maintains the postgres connection via piper, minimizing any
setup/teardown concerns)